Adds iAP2 BT/USB connection emulation over TCP transport
To be able to check iAP2 transport switch feature these adapters are
implemented. Implementation is done in the way of emulating iAP
transport over TCP since there is no available iAP library for open
source. Since it is just emulation not a real transport is will be built
only with other unit tests i.e. with BUILD_TESTS = ON